fix 2 part boolean search
