+++ package.json
@@ -16,6 +16,7 @@
     "core-js": "^3.6.5",
     "node-sass": "^6.0.1",
     "sass-loader": "10.2.0",
+    "stopword": "^1.0.11",
     "style-loader": "^3.3.1",
     "vue": "^2.6.14",
     "vue-feather-icons": "^5.1.0",
