+++ src/@core/libs/boolean_search.js
@@ -2,167 +2,7 @@
 /* eslint-disable no-unused-vars */
 import { FullTextSearch } from "@core/libs/search";
 import tinySegmenter from "@core/libs/tinySegmenter";
-
-
-class Expression {
-    isEmpty(arr) {
-        return arr.length === 0
-    }
-    // split string to operation 
-    toPostfix(exp) {
-        let postfix = [];
-        let stack = [];
-        let top = ""
-        let tokens = this.lex(exp);
-        tokens.forEach(token => {
-            if (this.isOperand(token)) {
-                postfix.push(token);
-            } else {
-                if (token === ")") {
-                    while ((top = stack.pop()) !== "(" && (!this.isEmpty(top))) {
-                        postfix.push(top);
-                    }
-                } else {
-                    while (
-                        (stack.length) && !(stack.at(-1) === "(") &&
-                        (this.priority(stack.at(-1)) >= this.priority(token))
-                    ) {
-                        postfix.push(stack.pop());
-                    }
-                    stack.push(token);
-                }
-            }
-        })
-        while (!this.isEmpty(stack)) {
-            postfix.push(stack.pop());
-        }
-        postfix = postfix.filter(x => x.trim().length > 0)
-        let operation = null, childOperation = null
-        console.log(postfix);
-        // if postfix not contain operation (~,|,&) or lenght =1
-        // return operation | with left is postfix string
-        if (postfix.length === 1 || postfix.find(x => this.isOperation(x)) === undefined) {
-            operation = {
-                operation: '|',
-                left: postfix.join(' '),
-            }
-        }
-        // create binary tree operation
-        // node:{
-        //     left: text or node,
-        //     right, text or node,
-        //     operatiop: |,&,~
-        // }
-        else {
-            for (let i = 0; i < postfix.length; i++) {
-                const element = postfix[i];
-                if (this.isOperator(element)) {
-                    operation = {
-                        right: childOperation,
-                        operation: element,
-                    }
-                    // if child operation is null , need find value of left and right 
-                    if (childOperation === null) {
-                        if (element === '~') {
-                            if (postfix.length > 1) {
-                                operation.left = postfix[i - 1]
-                                postfix.splice(i - 1, 2)
-                                i -= 2
-                            }
-                        }
-                        else {
-                            if (postfix.length > 2) {
-                                operation.right = postfix[i - 1]
-                                operation.left = postfix[i - 2]
-                                postfix.splice(i - 2, 3)
-                                i -= 3
-                            }
-                        }
-                    }
-                    // if  child operation not null , need find value of left ,right value is child operation
-                    else {
-                        if (element === '~') {
-                            if (postfix.length > 0) {
-                                operation.left = childOperation
-                                postfix.splice(i, 1)
-                                i -= 1
-                            }
-                        }
-                        else {
-                            if (postfix.length > 1) {
-                                operation.left = postfix[i - 1]
-                                postfix.splice(i - 1, 2)
-                                i -= 2
-                            }
-                        }
-                    }
-
-                    childOperation = operation
-                }
-            }
-        }
-        return operation;
-    }
-    // check none operation
-    isOperand(str) {
-        return !((str === "|") || (str === "&") || (str === "~") ||
-            (str === "(") || (str === ")"));
-    }
-
-    isOperator(str) {
-        return !this.isOperand(str);
-    }
-    isOperation(str) {
-        return (str === "|") || (str === "&") || (str === "~")
-    }
-
-    priority(operator) {
-        let priority = 0;
-
-        if (operator === ("&")) {
-            priority = 2;
-        }
-
-        if (operator === "~") {
-            priority = 3;
-        }
-
-        if (operator === "|") {
-            priority = 1;
-        }
-
-        if (operator === "(" || operator === ")") {
-            priority = 4;
-        }
-
-        return priority;
-    }
-
-    lex(string) {
-        let bad = [' or ', ' -', ' '];
-        let good = ['|', '~', '&'];
-        string = string.replace(bad, good);
-        string = string.toLowerCase();
-
-        let tokens = [];
-        let token = "";
-        for (let i = 0; i < string.length; i++) {
-            if (this.isOperator(string.charAt(i))) {
-                if (token) {
-                    tokens.push(token);
-                }
-                tokens.push(string.charAt(i));
-                token = "";
-            } else {
-                token = token + string.charAt(i);
-            }
-        }
-        if (token) {
-            tokens.push(token);
-        }
-        return tokens;
-    }
-}
+import { Expression } from './expression'
 
 
 export class BooleanSearch extends FullTextSearch {
@@ -183,8 +23,7 @@ export class BooleanSearch extends FullTextSearch {
             words.push(element.text)
         });
         let expression = new Expression;
-        let postfix = expression.toPostfix(query);
-        console.log(postfix);
+        let postfix = expression.createBinaryTree(query);
         let results = this.recursiveSearch(postfix)
         let questions = []
         results.forEach(element => {
@@ -225,7 +64,7 @@ export class BooleanSearch extends FullTextSearch {
     // if node is string=> return result by find text
     // if node is opreration =>return  recursive search
     getNodeResult(node) {
-        return this.isString(node) ? this.checkTextIncludeWords(this.splitText(node)) : this.recursiveSearch(node)
+        return this.isString(node) ? this.checkTextIncludeWords(tinySegmenter.removeStopWord(node)) : this.recursiveSearch(node)
     }
     isString(obj) {
         return (Object.prototype.toString.call(obj) === '[object String]');
