+++ src/const/index.js
@@ -1,0 +1,3 @@
+export default {
+    "SEARCH_BUTTON": false
+}
