+++ src/store/index.js
@@ -5,13 +5,15 @@ import Vuex from 'vuex'
 import auth from './modules/auth'
 import page from './modules/page'
 import category from "./modules/category"
+import config from "./modules/config"
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   modules: {
     auth,
     page,
-    category
+    category,
+    config
   },
   strict: process.env.DEV,
 })
