+++ src/store/modules/config/index.js
@@ -1,0 +1,16 @@
+import config from "@/const"
+export default {
+    namespaced: true,
+    state: {
+        data: config
+    },
+    getters: {
+
+    },
+    mutations: {
+
+    },
+    actions: {
+
+    }
+}
